osmanlica-tesseract/requirements.txt
# Tesseract OCR için Python bağımlılıkları
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0

# Görüntü ön işleme
scikit-image>=0.21.0
imutils>=0.5.4

# Veri işleme ve analiz
pandas>=2.0.0
matplotlib>=3.7.0

# Model eğitimi ve değerlendirme
scikit-learn>=1.3.0

# Yardımcı araçlar
tqdm>=4.65.0
click>=8.1.0
